pyproject.toml
[tool.poetry]
name = "tap-sailthru"
version = "0.0.1"
description = "`tap-sailthru` is a Singer tap for sailthru, built with the Meltano SDK for Singer Taps."
authors = ["Hunter Kuffel"]
keywords = [
    "ELT",
    "sailthru",
]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "<3.11,>=3.7.1"
requests = "^2.25.1"
singer-sdk = "^0.13.0"
sailthru-client = "^2.3.5"
google-cloud-bigquery = "^3.3.6"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
tox = "^3.24.4"
flake8 = "^3.9.2"
black = "^22.3.0"
pydocstyle = "^6.1.1"
mypy = "^0.910"
types-requests = "^2.26.1"
isort = "^5.10.1"

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "tap_sailthru"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-sailthru = 'tap_sailthru.tap:Tapsailthru.cli'
